refactor(object): extract Object.create prototype lowering
Move primitive prototype validation and the existing class-prototype fast path into the Object builtin subsystem module. This preserves #3768 behavior while keeping the static builtin driver below its LOC and per-function ceilings.
