+++ src/codegen/expressions/call-builtin-static.ts
@@ -94,6 +94,7 @@ import { compileStringLiteral } from "../string-ops.js";
 import { ensureStringRawHelper } from "../string-raw.js";
 import { defaultValueInstrs, pushDefaultValue } from "../type-coercion.js";
 import { compileMathCall } from "./builtins.js";
+import { tryCompileObjectCreateStaticPrototype } from "./call-object-builtins.js";
 import { emitLazyProtoGet } from "./extern.js";
 import { buildThrowJsErrorInstrs, emitThrowTypeError, noJsHost } from "./helpers.js";
 import { emitUndefined, ensureGetUndefined, ensureLateImport, flushLateImportShifts } from "./late-imports.js";
@@ -1927,45 +1928,8 @@ export function compileBuiltinStaticCall(
   ) {
     const arg0 = expr.arguments[0]!;
 
-    // ES5 §15.2.3.5 step 1: the requested [[Prototype]] must be an Object
-    // or null. The standalone native helper can only see an externref carrier;
-    // it historically treated every non-$Object carrier like null, so
-    // statically known primitive arguments silently created a null-prototype
-    // object instead of throwing. Preserve argument evaluation, then emit a
-    // catchable TypeError before entering the lenient native helper.
-    if (noJsHost(ctx)) {
-      const protoTag = ctx.oracle.staticJsTypeOf(arg0);
-      if (
-        protoTag === "number" ||
-        protoTag === "string" ||
-        protoTag === "boolean" ||
-        protoTag === "bigint" ||
-        protoTag === "symbol" ||
-        protoTag === "undefined"
-      ) {
-        const argType = compileExpression(ctx, fctx, arg0);
-        if (argType) fctx.body.push({ op: "drop" });
-        emitThrowTypeError(ctx, fctx, "Object prototype may only be an Object or null");
-        return { kind: "externref" };
-      }
-    }
-
-    // Object.create(Foo.prototype) → struct.new with default fields (Wasm-native fast path)
-    if (ts.isPropertyAccessExpression(arg0) && ts.isIdentifier(arg0.expression) && arg0.name.text === "prototype") {
-      const protoClassName = arg0.expression.text;
-      if (ctx.classSet.has(protoClassName)) {
-        const structTypeIdx = ctx.structMap.get(protoClassName);
-        const fields = ctx.structFields.get(protoClassName);
-        if (structTypeIdx !== undefined && fields) {
-          // Push default values for all fields, then struct.new
-          for (const field of fields) {
-            pushDefaultValue(fctx, field.type, ctx);
-          }
-          fctx.body.push({ op: "struct.new", typeIdx: structTypeIdx });
-          return { kind: "ref", typeIdx: structTypeIdx };
-        }
-      }
-    }
+    const staticPrototype = tryCompileObjectCreateStaticPrototype(ctx, fctx, arg0);
+    if (staticPrototype !== undefined) return staticPrototype;
 
     // Host import path: Object.create(null) and Object.create(proto[, descriptors])
     // Object.create(null) → empty object with null prototype
