+++ src/codegen/expressions/call-object-builtins.ts
@@ -10,8 +10,49 @@ import { resolveStoredObjectStaticMethod, resolveUncurriedObjectPrototypeMethod 
 import { compileObjectDefineProperties, compileObjectDefineProperty } from "../object-ops.js";
 import type { InnerResult } from "../shared.js";
 import { coerceType, compileExpression } from "../shared.js";
+import { pushDefaultValue } from "../type-coercion.js";
+import { emitThrowTypeError, noJsHost } from "./helpers.js";
 import { ensureLateImport, flushLateImportShifts } from "./late-imports.js";
 import { ts } from "../../ts-api.js";
+
+export function tryCompileObjectCreateStaticPrototype(
+  ctx: CodegenContext,
+  fctx: FunctionContext,
+  arg: ts.Expression,
+): InnerResult | undefined {
+  // ES5 §15.2.3.5 step 1: the requested [[Prototype]] must be an Object
+  // or null. The standalone native helper can only see an externref carrier;
+  // reject statically known primitives after preserving argument evaluation.
+  if (noJsHost(ctx)) {
+    const protoTag = ctx.oracle.staticJsTypeOf(arg);
+    if (
+      protoTag === "number" ||
+      protoTag === "string" ||
+      protoTag === "boolean" ||
+      protoTag === "bigint" ||
+      protoTag === "symbol" ||
+      protoTag === "undefined"
+    ) {
+      const argType = compileExpression(ctx, fctx, arg);
+      if (argType) fctx.body.push({ op: "drop" });
+      emitThrowTypeError(ctx, fctx, "Object prototype may only be an Object or null");
+      return { kind: "externref" };
+    }
+  }
+
+  // Object.create(Foo.prototype) → struct.new with default fields (Wasm-native fast path)
+  if (ts.isPropertyAccessExpression(arg) && ts.isIdentifier(arg.expression) && arg.name.text === "prototype") {
+    const protoClassName = arg.expression.text;
+    const structTypeIdx = ctx.structMap.get(protoClassName);
+    const fields = ctx.structFields.get(protoClassName);
+    if (ctx.classSet.has(protoClassName) && structTypeIdx !== undefined && fields) {
+      for (const field of fields) pushDefaultValue(fctx, field.type, ctx);
+      fctx.body.push({ op: "struct.new", typeIdx: structTypeIdx });
+      return { kind: "ref", typeIdx: structTypeIdx };
+    }
+  }
+  return undefined;
+}
 
 function emitExternrefArgument(
   ctx: CodegenContext,
